Move fuzz main function to fuzz_main.cpp
